+++ app/Console/Commands/AngularFeature.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
+use File;
 
 class AngularFeature extends Command{
 
@@ -32,10 +33,14 @@ class AngularFeature extends Command{
      */
     public function handle(){
         $name = $this->argument('name');
+        $studly_name = studly_case($name);
 
         $html = file_get_contents(__DIR__ . '/Stubs/AngularFeature/feature.html.stub');
-        $js = file_get_contents(__DIR__ . '/Stubs/AngularFeature/feature.js.stub');
+        $js   = file_get_contents(__DIR__ . '/Stubs/AngularFeature/feature.js.stub');
         $less = file_get_contents(__DIR__ . '/Stubs/AngularFeature/feature.less.stub');
+
+        $html = str_replace('{{StudlyName}}', $studly_name, $html);
+        $js = str_replace('{{StudlyName}}', $studly_name, $js);
 
         $folder = __DIR__ . '/../../../angular/app/' . $name;
         if (is_dir($folder)){
@@ -45,15 +50,17 @@ class AngularFeature extends Command{
         }
 
         //create folder
-        \File::makeDirectory($folder, 0775, true);
+        File::makeDirectory($folder, 0775, true);
 
         //create view (.html)
-        \File::put($folder . '/' . $name . '.html', $html);
+        File::put($folder . '/' . $name . '.html', $html);
 
         //create controller (.js)
-        \File::put($folder . '/' . $name . '.js', $js);
+        File::put($folder . '/' . $name . '.js', $js);
 
         //create less file (.less)
         \File::put($folder . '/' . $name . '.less', $less);
+
+        $this->info('Feature created successfully.');
     }
 }