+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### v2.6.2
++ fixed elixir watcher for less
+
 ### v2.6.1
 + updated bower dependencies (angular material released 0.10.1)
 + added artisan generator for a new feature: **artisan ng:feature**